requirements.txt
numpy==1.26.0
requests==2.31.0
Pillow==10.0.0
matplotlib==3.8.0
seaborn==0.12.2
scikit-learn==1.2.2